Fix compiler warnings (#539)
* fix: resolve non-index compiler warnings

* Fix a few other warnings

* Fix more warnings

* Fix more

* Fix more

* Do not check for tracks_begin being >= 0

* Make a few more invasive changes

* style: apply clang-format

* fix: suppress FastJet deprecated-copy warnings

---------

+++ analyzers/dataframe/src/JetFlavourUtils.cc
@@ -1,6 +1,7 @@
 #include "FCCAnalyses/JetFlavourUtils.h"
 #include "ONNXRuntime/WeaverInterface.h"
 
+#include <cstddef>
 #include <memory>
 
 namespace FCCAnalyses {
@@ -49,9 +50,9 @@ namespace FCCAnalyses {
         throw std::runtime_error("Invalid index requested for jet flavour weight.");
       rv::RVec<float> out;
       for (const auto& jet_weights : jets_weights) {
-        if (weight >= jet_weights.size())
+        if (static_cast<std::size_t>(weight) >= jet_weights.size())
           throw std::runtime_error("Flavour weight index exceeds the number of weights registered.");
-        out.emplace_back(jet_weights.at(weight));
+        out.emplace_back(jet_weights.at(static_cast<std::size_t>(weight)));
       }
       return out;
     }
